requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-allauth==0.57.0
django-allauth-2fa==0.11.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
PyJWT==2.8.0
python-decouple==3.8
cryptography==41.0.7
bcrypt==4.0.1
phonenumbers==8.13.27
django-otp==1.3.0
qrcode==7.4.2

# Payments
stripe==7.5.0
web3==6.11.1
bitcoinlib==0.6.14
coinbase==2.1.0

# Email & Notifications
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
sendgrid==6.10.0
twilio==8.10.0
channels==4.0.0
channels-redis==4.1.0

# PDF & Documents
reportlab==4.0.9
weasyprint==60.1
xhtml2pdf==0.2.12
qrcode[pil]==7.4.2

# Images & Media
Pillow==10.1.0
django-imagekit==5.0.0
easy-thumbnails==2.8.5

# API & Utilities
django-rest-knox==4.2.0
drf-yasg==1.21.7
django-import-export==3.3.1
django-admin-interface==0.19.1
django-advanced-filters==1.3.0

# Performance
django-debug-toolbar==4.2.0
django-storages==1.14.2
boto3==1.34.1
django-redis==5.4.0

# Monitoring
sentry-sdk==1.38.0
django-prometheus==2.3.1
prometheus-client==0.19.0

# Development
python-dotenv==1.0.0
faker==20.1.0
factory-boy==3.3.0
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.0
flake8==7.0.0
pre-commit==3.6.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0